% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corrplot.R
\name{get_corrplot}
\alias{get_corrplot}
\title{get_corrplot}
\usage{
get_corrplot(nba, year, method = "number")
}
\arguments{
\item{nba}{the database of nba players}

\item{year}{the year you are looking at}

\item{method}{the type of correlogram to display}
}
\description{
This function plots the correlogram of numeric continuous variables only for a given year.
}
\examples{
get_corrplot(nba,1950)
}
\keyword{correlogram}
