% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_older_player.R
\name{get_older_player}
\alias{get_older_player}
\title{get_older_player

This function returns the older player for a given year. (for two players with the same age, alphabetical order is used)}
\usage{
get_older_player(nba, year)
}
\arguments{
\item{nba}{the database of nba players}

\item{year}{the year you are looking at}
}
\description{
get_older_player

This function returns the older player for a given year. (for two players with the same age, alphabetical order is used)
}
\examples{
get_older_player(nba,1950)
}
\keyword{older}
\keyword{player}
