% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_point_interest.R
\name{get_point_interest}
\alias{get_point_interest}
\title{get_point_interest function}
\usage{
get_point_interest(nba, year, player, col_name)
}
\arguments{
\item{nba}{the database of nba players}

\item{year}{the year you are looking at}

\item{player}{the player you are looking for}

\item{col_name}{the data you want}
}
\description{
This function allows you return a data point of interest in the dataset for some player given an individual year.
}
\examples{
get_point_interest(nba,1950,"Curly Armstrong",Pos)
}
\keyword{interest}
\keyword{of}
\keyword{point}
