% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba.R
\docType{data}
\name{nba}
\alias{nba}
\title{NBA dataset}
\description{
The data-set contains aggregate individual statistics for 67 NBA seasons. from basic box-score attributes such as points, assists, rebounds etc., to more advanced money-ball like features such as Value Over Replacement.
}
\examples{
View(nba)
}
\keyword{data,}
\keyword{nba}
